Clean up: add new example for untyped handoff
